[hotfix][test] Introduce InputGateTestBase and deduplicate test code
